change the loopFindNearKeyframesCloud function to be more clear
The function is also changed in the author's latest code base
https://github.com/gisbi-kim/lt-mapper/blob/bce4236d97bf0d9d97187f9355380e981ac7e499/ltslam/src/LTslam.cpp#L201

Basically loopFindNearKeyframesCloud function does not require the fourth input parameter that it initially had - it just creates more confusion and it never is used. Also while transforming the pointcloud to global coordinate frame (origin) and to accumulate surrounding frames, one does not need the fourth input parameter. Changed the function accordingly.